pyproject.toml
[tool.poetry]
name = "gsheet"
version = "0.1.0"
description = ""
authors = ["sth-v <aa@contextmachine.online>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httplib2 = "^0.22.0"
google-api-python-client="*"
google-auth="*"
python-dotenv="*"
fastapi={version="*", extras=['all']}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
